Split prewarm and rendering-state code out of the over-cap files
workflow-guard-tests enforces a 900-line hard cap: a PR may not grow any
file already over it, and both BrowserPanel.swift (+39) and
BrowserWindowPortal.swift (+14) grew. Comply by splitting instead of
bumping budgets:

- Sources/Panels/BrowserNavigationPolicy.swift: the insecure-HTTP and
  navigation-request free functions move out of BrowserPanel.swift.
- Sources/Panels/BrowserPanel+PrewarmedWebViewAdoption.swift:
  resolvedProfileID and the new claimedPrewarmedWebView eligibility
  gate; the initializer now makes one call. makeWebView becomes
  internal so the pool builds identically-configured webviews directly.
- Sources/BrowserPortalWebViewRenderingState.swift: the WKWebView
  hidden/reattach rendering-state cluster (flag, notify-hidden,
  reattach selectors, adoption wrapper) moves out of
  BrowserWindowPortal.swift.

Both files now end below their merge-base line counts (11641 vs 11669
and 4036 vs 4121) and the budget entries are set to the new actuals.
