+++ Sources/Panels/BrowserNavigationPolicy.swift
@@ -1,0 +1,53 @@
+import AppKit
+import Foundation
+
+/// Free-function navigation policy shared by browser panels, popups, and the
+/// navigation delegate, extracted from BrowserPanel.swift.
+
+func browserShouldBlockInsecureHTTPURL(
+    _ url: URL,
+    defaults: UserDefaults = .standard
+) -> Bool {
+    browserShouldBlockInsecureHTTPURL(
+        url,
+        rawAllowlist: defaults.string(forKey: BrowserInsecureHTTPSettings.allowlistKey)
+    )
+}
+
+func browserShouldBlockInsecureHTTPURL(
+    _ url: URL,
+    rawAllowlist: String?
+) -> Bool {
+    guard url.scheme?.lowercased() == "http" else { return false }
+    guard let host = BrowserInsecureHTTPSettings.normalizeHost(url.host ?? "") else { return true }
+    return !BrowserInsecureHTTPSettings.isHostAllowed(host, rawAllowlist: rawAllowlist)
+}
+
+func browserShouldConsumeOneTimeInsecureHTTPBypass(
+    _ url: URL,
+    bypassHostOnce: inout String?
+) -> Bool {
+    guard let bypassHost = bypassHostOnce else { return false }
+    guard url.scheme?.lowercased() == "http",
+          let host = BrowserInsecureHTTPSettings.normalizeHost(url.host ?? "") else {
+        return false
+    }
+    guard host == bypassHost else { return false }
+    bypassHostOnce = nil
+    return true
+}
+
+func browserShouldPersistInsecureHTTPAllowlistSelection(
+    response: NSApplication.ModalResponse,
+    suppressionEnabled: Bool
+) -> Bool {
+    guard suppressionEnabled else { return false }
+    return response == .alertFirstButtonReturn || response == .alertSecondButtonReturn
+}
+
+func browserPreparedNavigationRequest(_ request: URLRequest) -> URLRequest {
+    var preparedRequest = request
+    // Match browser behavior for ordinary loads while preserving method/body/headers.
+    preparedRequest.cachePolicy = .useProtocolCachePolicy
+    return preparedRequest
+}
